refactor: Players validater 클래스 추출 및 허용인원수 enum 클래스 분리
